Restrict exercise definition controller to authorization

+++ Controllers/ExerciseDefinitionController.cs
@@ -4,9 +4,12 @@ using Microsoft.AspNetCore.Mvc;
 using RobicServer.Services;
 using RobicServer.Models;
 using System.Threading.Tasks;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 
 namespace RobicServer.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class ExerciseDefinitionController : ControllerBase
@@ -19,22 +22,31 @@ namespace RobicServer.Controllers
         }
 
         [HttpGet]
-        public List<ExerciseDefiniton> Get() => _exerciseDefintionRepo.AsQueryable().ToList();
+        public List<ExerciseDefiniton> Get()
+        {
+            string userId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            return _exerciseDefintionRepo.AsQueryable().Where(exercise => exercise.User == userId).ToList();
+        }
 
         [HttpGet("{id:length(24)}", Name = "GetExerciseDefinition")]
         public async Task<IActionResult> Get(string id)
         {
             ExerciseDefiniton exercise = await _exerciseDefintionRepo.FindByIdAsync(id);
             if (exercise == null)
-            {
                 return NotFound();
-            }
+
+            if (exercise.User != User.FindFirst(ClaimTypes.NameIdentifier).Value)
+                return Unauthorized();
+
             return Ok(exercise);
         }
 
         [HttpPost]
         public async Task<IActionResult> Create(ExerciseDefiniton exercise)
         {
+            if (exercise.User != User.FindFirst(ClaimTypes.NameIdentifier).Value)
+                return Unauthorized();
+
             await _exerciseDefintionRepo.InsertOneAsync(exercise);
             return CreatedAtRoute("GetExerciseDefinition", new { id = exercise.Id }, exercise);
         }
@@ -44,9 +56,10 @@ namespace RobicServer.Controllers
         {
             ExerciseDefiniton exercise = await _exerciseDefintionRepo.FindByIdAsync(id);
             if (exercise == null)
-            {
                 return NotFound();
-            }
+
+            if (exercise.User != User.FindFirst(ClaimTypes.NameIdentifier).Value)
+                return Unauthorized();
 
             await _exerciseDefintionRepo.ReplaceOneAsync(updatedExercise);
             return Ok(updatedExercise);
@@ -57,9 +70,10 @@ namespace RobicServer.Controllers
         {
             ExerciseDefiniton exercise = await _exerciseDefintionRepo.FindByIdAsync(id);
             if (exercise == null)
-            {
                 return NotFound();
-            }
+
+            if (exercise.User != User.FindFirst(ClaimTypes.NameIdentifier).Value)
+                return Unauthorized();
 
             await _exerciseDefintionRepo.DeleteByIdAsync(id);
             return NoContent();
